requirements.txt
flask
numpy
matplotlib
pillow
sympy
plotly
imageio
Flask-Bootstrap
seaborn